refactor: group trace files in new folder
